refactor: restructure app folders and split main
