+++ lib/constants/api_keys.dart
@@ -1,0 +1,2 @@
+/// API keys used by the application.
+const String placesWebApiKeyB = 'AIzaSyCsxalL8q8DYpCm3wfyuU_y9yAenj6Mifw';
