+++ lib/models/scored_place.dart
@@ -1,0 +1,8 @@
+import 'place_item.dart';
+
+class ScoredPlace {
+  final PlaceItem item;
+  final double score;
+
+  ScoredPlace(this.item, this.score);
+}
